add authorization and authentication with jwt

+++ routers/todos.py
@@ -2,32 +2,28 @@ from typing import Annotated, Optional
 from fastapi import APIRouter, Depends, status, Path, HTTPException
 from pydantic import Field, field_validator, BaseModel
 import models
-from database import SessionLocal
+from database import get_db
 from sqlalchemy.orm import Session
+from .auth import get_current_user
 
 router = APIRouter()
 
-
-def get_db():
-    db = SessionLocal()
-    try:
-        yield db
-    finally:
-        db.close()
-
-
 db_dependency = Annotated[Session, Depends(get_db)]
+user_dependency = Annotated[dict, Depends(get_current_user)]
 
 
 @router.get("/", status_code=status.HTTP_200_OK)
-async def read_all(db: db_dependency):
-    return db.query(models.Todos).all()
+async def read_all(user: user_dependency, db: db_dependency):
+    if user is None:
+        raise HTTPException(status_code=status.HTTP_401_UNAUTHORIZED, detail="Unauthorized")
+    return db.query(models.Todos).filter(models.Todos.owner == user.get("id")).all()
 
 
 @router.get("/todo/{todo_id}", status_code=status.HTTP_200_OK)
-async def read_todo(db: db_dependency, todo_id: int = Path(gt=0)):
-    todo_model = db.query(models.Todos).filter(
-        models.Todos.id == todo_id).first()  # save some time by adding .first()
+async def read_todo(user: user_dependency, db: db_dependency, todo_id: int = Path(gt=0)):
+    if user is None:
+        raise HTTPException(status_code=status.HTTP_401_UNAUTHORIZED, detail="Unauthorized")
+    todo_model = db.query(models.Todos).filter(models.Todos.id == todo_id).filter(models.Todos.owner == user.get("id")).first()  # save some time by adding .first()
     if todo_model is not None:
         return todo_model
     else:
@@ -48,13 +44,15 @@ class TodoRequest(BaseModel):
 
 
 @router.post("/todo", status_code=status.HTTP_200_OK)
-async def create_todo(db: db_dependency, todo_request: TodoRequest):
-    new_todo = models.Todos(**todo_request.model_dump())
+async def create_todo(user: user_dependency, db: db_dependency, todo_request: TodoRequest):
+    if user is None:
+        raise HTTPException(status_code=status.HTTP_401_UNAUTHORIZED, detail="Unauthorized")
+    new_todo = models.Todos(**todo_request.model_dump(), owner=user.get("id"))
     db.add(new_todo)
     db.commit()
 
     added_todo = db.query(models.Todos).filter_by(
-        title=new_todo.title, description=new_todo.description, priority=new_todo.priority, complete=new_todo.complete).first()
+        title=new_todo.title, description=new_todo.description, priority=new_todo.priority, complete=new_todo.complete, owner=new_todo.owner).first()
     if added_todo:
         return added_todo
     else:
@@ -62,8 +60,11 @@ async def create_todo(db: db_dependency, todo_request: TodoRequest):
 
 
 @router.put("/todo/{todo_id}", status_code=status.HTTP_204_NO_CONTENT)
-async def update_todo(db: db_dependency, todo_request: TodoRequest, todo_id: int = Path(gt=0)):
-    todo = db.query(models.Todos).filter_by(id=todo_id).first()
+async def update_todo(user: user_dependency, db: db_dependency, todo_request: TodoRequest, todo_id: int = Path(gt=0)):
+    if user is None:
+        raise HTTPException(status_code=status.HTTP_401_UNAUTHORIZED, detail="Unauthorized")
+    
+    todo = db.query(models.Todos).filter_by(id=todo_id, owner=user.get("id")).first()
     if todo is None:
         raise HTTPException(status_code=404, detail="todo not found")
 
@@ -89,8 +90,11 @@ class PatchRequest(BaseModel):
 
 
 @router.patch("/todo/{todo_id}", status_code=status.HTTP_200_OK)
-async def patch_todo(db: db_dependency, todo_request: PatchRequest, todo_id: int = Path(gt=0)):
-    todo = db.query(models.Todos).filter(models.Todos.id == todo_id).first()
+async def patch_todo(user: user_dependency,db: db_dependency, todo_request: PatchRequest, todo_id: int = Path(gt=0)):
+    if user is None:
+        raise HTTPException(status_code=status.HTTP_401_UNAUTHORIZED, detail="Unauthorized")
+    
+    todo = db.query(models.Todos).filter(models.Todos.id == todo_id, models.Todos.owner == user.get("id")).first()
     if todo is None:
         raise HTTPException(status_code=404, detail="todo not found")
 
@@ -100,13 +104,25 @@ async def patch_todo(db: db_dependency, todo_request: PatchRequest, todo_id: int
     todo.complete = todo.complete if todo_request.complete is None else todo_request.complete
 
     db.commit()
+    db.refresh(todo) # we need to refresh todo to access the updated data
     return todo
 
 
 @router.delete("/todo/{todo_id}", status_code=status.HTTP_204_NO_CONTENT)
-async def delete_todo(db: db_dependency, todo_id: int = Path(gt=0)):
-    todo = db.query(models.Todos).filter(models.Todos.id == todo_id).first()
+async def delete_todo(user: user_dependency, db: db_dependency, todo_id: int = Path(gt=0)):
+    if user is None:
+        raise HTTPException(status_code=status.HTTP_401_UNAUTHORIZED, detail="Unauthorized")
+    
+    todo = db.query(models.Todos).filter((models.Todos.id == todo_id) & (models.Todos.owner == user.get("id"))).first()
     if todo is None:
         raise HTTPException(status_code=404, detail="todo not found")
-    db.delete(todo)
+    
+    db.delete(todo)  # recommended delete method. the following two methods remove the relationship between todo and user, if cascade = "all, delete-orphan" was not claimed, the removed todo will still be in Todos table, just without an owner
+
+    '''
+    db_user = db.query(models.Users).filter_by(id=user.get("id")).first()
+    db_user.todos.remove(todo)
+    '''
+
+    # todo.user.todos.remove(todo)
     db.commit()
